Canvas with default objects and transformations

+++ src/main/java/com/example/demo/entities/Color3.java
@@ -1,11 +1,11 @@
 package com.example.demo.entities;
 
 public class Color3 {
-    private float red;    // Cor vermelha
-    private float green;  // Cor verde
-    private float blue;   // Cor azul
+    private double red;    // Cor vermelha
+    private double green;  // Cor verde
+    private double blue;   // Cor azul
 
-    public Color3(float red, float green, float blue) {
+    public Color3(double red, double green, double blue) {
         // Valores entre [0.0, 1.0]
         this.red = Math.min(Math.max(red, 0.0f), 1.0f);
         this.green = Math.min(Math.max(green, 0.0f), 1.0f);
@@ -25,27 +25,27 @@ public class Color3 {
     }
 
     // Métodos para obter os componentes RGB
-    public float getRed() {
+    public double getRed() {
         return red;
     }
 
-    public float getGreen() {
+    public double getGreen() {
         return green;
     }
 
-    public float getBlue() {
+    public double getBlue() {
         return blue;
     }
 
-    public void setRed(float red) {
+    public void setRed(double red) {
         this.red = red;
     }
 
-    public void setGreen(float green) {
+    public void setGreen(double green) {
         this.green = green;
     }
 
-    public void setBlue(float blue) {
+    public void setBlue(double blue) {
         this.blue = blue;
     }
 }
